reorganize shared crate to be scoped only to adl
