+++ rust/adl/src/custom/common/db.rs
