+++ rust/adl/src/custom/common/mod.rs
