+++ rust/adl/src/gen/common/db.rs
