+++ rust/adl/src/gen/common/http.rs
