+++ rust/adl/src/gen/common/mod.rs
