+++ rust/adl/src/gen/common/strings.rs
