+++ rust/adl/src/gen/protoapp/apis/mod.rs
