+++ rust/server/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-shared = { workspace = true }
+adl = { workspace = true }
 anyhow = "1.0.70"
 async-trait = "0.1.68"
 base64 = "0.21.2"
