+++ rust/server/src/bin/protoapp-server.rs
@@ -1,7 +1,7 @@
 use clap::Parser;
 use std::path::PathBuf;
 
-use shared::adl::gen::protoapp::config::server::ServerConfig;
+use adl::gen::protoapp::config::server::ServerConfig;
 
 use protoapp::server;
